// src/__generated__/pageProductsPaginationFragment.graphql.ts
export type pageProductsPaginationFragment$data = {
  readonly productsConnection: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly category: string;
        readonly createdAt: any;
        readonly description: string;
        readonly id: string;
        readonly inventory: number;
        readonly isActive: boolean;
        readonly name: string;
        readonly photoUrl: string;
        readonly price: number;
        readonly slug: string;
        readonly tags: ReadonlyArray<string>;
        readonly updatedAt: any;
        readonly videoUrl: string | null;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null;
    };
  };
  readonly " $fragmentType": "pageProductsPaginationFragment";
};
export type pageProductsPaginationFragment$key = {
  readonly " $data"?: pageProductsPaginationFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"pageProductsPaginationFragment">;
};